+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Version 2.10 (not yet released)
+## Version 2.10
 
 * Added clone butten to the edit rule screen, so you can copy and change only what needs changing (#29)
 * Added configuration for the record type for Activity and Case Activity trigger.